requirements.txt
streamlit==1.36.0
pandas
openpyxl
PyPDF2
python-docx